requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
binpacking==1.5.2
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.2
cpm-kernels==1.0.11
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
dacite==1.9.2
datasets==3.6.0
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
emoji==2.15.0
exceptiongroup==1.3.1
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
future==1.0.0
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.2
json_repair==0.54.2
kiwisolver==1.4.9
langdetect==1.0.9
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
mpmath==1.3.0
ms_swift==3.11.1
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.15
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.11.0
orjson==3.11.5
oss2==2.19.1
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==11.3.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge==1.0.1
ruff==0.14.9
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
syllapy==0.7.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
transformers-stream-generator==0.0.5
triton==3.5.1
trl==0.24.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
